Switch to CLI wrapper
Copies the CLI wrapper structure from houndci/swift and repurposes it to
work with scss-lint. No major modifications were made.

This is a first step to potentially moving all of Hound's services to a
CLI wrapper approach, with the primary goal of increasing stability and
using the linters in their natural habitat.