Report haml-lint errors as violations
When haml-lint reports an error, e.g. invalid syntax, it does not
include a rule name in the message. The regex for parsing these messages
is assuming a rule name will be present and is failing to capture error
messages because of it.

Additionally, when reporting errors, haml-lint seems to use zero-based
indexing.

This commit:

* Updates `HamlLint::Violation` regex pattern to have an optional rule
  name, so errors are matched
* Updates `HamlLint::Violation#line_number` to increment the line number
  by one when the message is an error, so the correct line number is
  reported

Related: https://github.com/houndci/hound/pull/1112