Refactor to simplify adding more linters
To add additional linter, it is now possible to copy one of the existing
linter directories (e.b. `lib/linters/scss_lint`) and modify the
`Options` for linter specific command and config, as well as `Tokenizer`
for system output parsing specifics.

This is an attempt to minimize the moving pieces for adding a new linter.

In linter specific `Options`:
  * Linter command to run
  * Linter's default config file (Hound defaults)
  * Config filename which the linter expects by default

In linter specific `Tokenizer`:
  * Parsing logic to extract violations from command output