Add ESLint
This is using a ruby wrapper for ESLint.

This will not ignore files natively, since ignore file isn't sent as
part of the job payload.

Hound proper handles ignoring JS files, and does not schedule jobs if
files are ignored. So, we are okay for now.
At some point we might want to consider supporting native file ignores,
but then we'll be scheduling jobs for files that are simply ignored.