Set default Ruby version for RuboCop
Otherwise, RuboCop will assume an oldest supported Ruby version if it's
not specified by the user. From their docs:
```
What MRI version of the Ruby interpreter is the inspected code intended
to run on? (If there is more than one, set this to the lowest version.)
If a value is specified for TargetRubyVersion then it is used.
Else if .ruby-version exists and it contains an MRI version it is used.
Otherwise we fallback to the oldest officially supported Ruby version
(2.0).
```

Thus, we should either default it to `2.3`, or tell users that get
strange "token" (code parse) errors to specify `TargetRubyVersion` in
their custom config.