Add Credo linter to lint Elixir files
Add a job for the Credo linter to support linting elixir files.

Use `asdf` for Elixir/Erlang versioning and configure CircleCI to install
them. Since the `mix` command isn't run in the directory of the project, we need
to configure the global versions for `asdf` which we do by copying the
`.tool-versions` to the home directory.

Running mix outside of the directory with the `mix.exs` file isn't really
supported well, so we need to do two things to get it to work. Pass in the
`MIX_EXS` environment variable and make sure the `mix.exs` file uses absolute
paths instead of relative ones since `mix` doesn't expand them automatically.

Update setup script remote addition to hide output and not care if it passes or
not. If it already exists it exists with a non-zero status code and makes CI
think this script fails. The `|| true` takes care of that.

Add buildpack config for Heroku