Add link to code smell explanation
When reek finds a code smell it will provide a link to the project's
wiki where an explanation for that code smell is. There is further
information about what it's "smelling", how it may be configured (on or
off) and what it's checking for.

This commit provides that link, but formatted with markdown in order to
reduce noise in the comment.