Add `sass-lint` to linters (#112)
This commit adds `sass-lint` to the available linters for Hound by
adding the package to the list of yarn dependencies and creating the
associated `Options` and `Tokenizer` objects.

Because the linter is written in pure Node, its command isn't available
in environments that haven't run a `yarn global add` (like CircleCI). In
light of that, the `SassLint::Options.command` builds by pointing to the
file in the `node-modules` directory. The command also has 2 flags.
The `-q` flag prevents sass-lint's CLI from exiting as soon as it hits
an error (severity of `2` in its config file), and the `-v` makes it
report warnings (severity of `1`).